Rozaq

+++ application/controllers/c_ktgrsampah.php
@@ -17,7 +17,7 @@ class c_ktgrsampah extends CI_Controller {
 
         $user_id = $this->session->userdata('user_id');
         $data['user'] = $this->m_users->get_user_by_id($user_id);
-        $data['ktgrsampah'] = $this->m_ktgrsampah->get_all_ktgrsampah();
+        $data['ktgrsampah'] = $this->m_ktgrsampah->get_all_ktgrsampah($user_id);
 
         $this->load->view('templates/header', $data);
         $this->load->view('templates/topbar', $data);
@@ -53,10 +53,12 @@ class c_ktgrsampah extends CI_Controller {
         if ($this->form_validation->run() == FALSE) {
             $this->create();
         } else {
+            $user_id = $this->session->userdata('user_id');
             $data = array(
                 'nama_kategori' => $this->input->post('nama_kategori'),
                 'deskripsi' => $this->input->post('deskripsi'),
-                'warna_kategori' => $this->input->post('warna_kategori')
+                'warna_kategori' => $this->input->post('warna_kategori'),
+                'user_id' => $user_id
             );
 
             $this->m_ktgrsampah->insert_ktgrsampah($data);
@@ -72,6 +74,11 @@ class c_ktgrsampah extends CI_Controller {
         $user_id = $this->session->userdata('user_id');
         $data['user'] = $this->m_users->get_user_by_id($user_id);
         $data['ktgrsampah'] = $this->m_ktgrsampah->get_ktgrsampah($id_ktgrsampah);
+
+        if ($data['ktgrsampah']['user_id'] != $user_id) {
+            // Handle unauthorized access (e.g., redirect or show error)
+            redirect('c_ktgrsampah');
+        }
 
         $this->load->view('templates/header', $data);
         $this->load->view('templates/topbar', $data);
@@ -92,6 +99,7 @@ class c_ktgrsampah extends CI_Controller {
         if ($this->form_validation->run() == FALSE) {
             $this->edit($id_ktgrsampah);
         } else {
+            $user_id = $this->session->userdata('user_id');
             $data = array(
                 'nama_kategori' => $this->input->post('nama_kategori'),
                 'deskripsi' => $this->input->post('deskripsi'),
@@ -108,7 +116,13 @@ class c_ktgrsampah extends CI_Controller {
             redirect('auth/login');
         }
 
-        $this->m_ktgrsampah->delete_ktgrsampah($id_ktgrsampah);
+        $user_id = $this->session->userdata('user_id');
+        $ktgrsampah = $this->m_ktgrsampah->get_ktgrsampah($id_ktgrsampah);
+
+        if ($ktgrsampah['user_id'] == $user_id) {
+            $this->m_ktgrsampah->delete_ktgrsampah($id_ktgrsampah);
+        }
+        
         redirect('c_ktgrsampah');
     }
 }
